requirements.txt
# LangGraph and dependencies
langgraph>=0.0.20
langgraph-checkpoint>=0.0.1
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# API clients
openai>=1.0.0
anthropic>=0.8.0

# TTS
piper-tts>=0.1.0
TTS>=0.22.0

# Whisper
faster-whisper>=0.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Video processing
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Utilities
python-dotenv>=1.0.0
